Move run_splinter, run_sawtooth into splinter/sawtooth modules
This drastically cuts down on the cfg statements required, as most of
the cfg guards were for splinter or sawtooth support.
