+++ daemon/src/splinter/mod.rs
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Cargill Incorporated
+ * Copyright 2019-2021 Cargill Incorporated
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,3 +19,6 @@ pub mod app_auth_handler;
 pub mod batch_submitter;
 pub mod event;
 pub mod key;
+mod run;
+
+pub use run::run_splinter;
